- fixed bug in Game.addPlayers() where count was not being cast to int - updated tests for Game.addPlayers() to test count returned as well

+++ src/models/game.ts
@@ -236,7 +236,7 @@ class Game {
     }
 
     const result: QueryResult<CountResultInterface> = await db.query(`
-        SELECT COUNT(*)
+        SELECT COUNT(*)::int
         FROM game_players
         WHERE game_id = $1
     `, [gameId]);
